requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[argon2]==1.7.4
python-dotenv==1.0.0
redis==5.0.1
aioredis==2.0.1
httpx==0.25.1
psycopg2-binary==2.9.9